+++ internal/services/admin.go
@@ -124,7 +124,34 @@ func (s AdminService) GetSettings(
 		platforms = &count
 	}
 
-	return models.NewAdminSettingsResponse(s.Config, platforms, s.workerCoverage(logger)), nil
+	_, deletionQueued := s.Config.Events.Queues[configuration.EventsObjectDeletion]
+	_, bucketQueued := s.Config.Events.Queues[configuration.EventsBucketEvents]
+	confirmsUploads := configuration.RequiresUploadConfirmation(s.Config.Storage.Type, s.Config.Events.Type)
+
+	status := func(name string, applicable bool) models.CoverageStatus {
+		if !applicable {
+			return models.CoverageNotApplicable
+		}
+		covered, err := cache.IsWorkerCovered(s.Cache, name)
+		if err != nil {
+			logger.Error("Failed to check worker coverage", zap.String("worker", name), zap.Error(err))
+			return models.CoverageUnknown
+		}
+		if covered {
+			return models.CoverageCovered
+		}
+		return models.CoverageNotCovered
+	}
+
+	coverage := models.WorkerSettings{
+		HTTPServer:       status(configuration.CoverageHTTPServer, true),
+		ObjectDeletion:   status(configuration.WorkerObjectDeletion, deletionQueued),
+		BucketEvents:     status(configuration.WorkerBucketEvents, bucketQueued),
+		TrashCleanup:     status(configuration.WorkerTrashCleanup, confirmsUploads),
+		GarbageCollector: status(configuration.WorkerGarbageCollector, true),
+	}
+
+	return models.NewAdminSettingsResponse(s.Config, platforms, coverage), nil
 }
 
 func (s AdminService) GetBucketList(
@@ -174,33 +201,4 @@ func (s AdminService) GetBucketList(
 	}
 
 	return result
-}
-
-func (s AdminService) workerCoverage(logger *zap.Logger) models.WorkerSettings {
-	_, deletionQueued := s.Config.Events.Queues[configuration.EventsObjectDeletion]
-	_, bucketQueued := s.Config.Events.Queues[configuration.EventsBucketEvents]
-	confirmsUploads := configuration.RequiresUploadConfirmation(s.Config.Storage.Type, s.Config.Events.Type)
-
-	status := func(name string, applicable bool) models.CoverageStatus {
-		if !applicable {
-			return models.CoverageNotApplicable
-		}
-		covered, err := cache.IsWorkerCovered(s.Cache, name)
-		if err != nil {
-			logger.Error("Failed to check worker coverage", zap.String("worker", name), zap.Error(err))
-			return models.CoverageUnknown
-		}
-		if covered {
-			return models.CoverageCovered
-		}
-		return models.CoverageNotCovered
-	}
-
-	return models.WorkerSettings{
-		HTTPServer:       status(configuration.CoverageHTTPServer, true),
-		ObjectDeletion:   status(configuration.WorkerObjectDeletion, deletionQueued),
-		BucketEvents:     status(configuration.WorkerBucketEvents, bucketQueued),
-		TrashCleanup:     status(configuration.WorkerTrashCleanup, confirmsUploads),
-		GarbageCollector: status(configuration.WorkerGarbageCollector, true),
-	}
 }
