+++ web/src/locales/fr.json
@@ -893,8 +893,8 @@
         "path": "Chemin",
         "tls_server_name": "Nom du serveur TLS",
         "bucket_name": "Bucket",
-        "endpoint": "Point de terminaison",
-        "external_endpoint": "Point de terminaison externe",
+        "endpoint": "Point d'exposition",
+        "external_endpoint": "Point d'exposition externe",
         "region": "Région",
         "project_id": "ID du projet",
         "force_path_style": "Forcer le path-style",
